+++ src/loading/cob/sections/mod.rs
@@ -3,11 +3,9 @@ mod cob_defs;
 mod cob_import;
 mod cob_manifest;
 mod cob_scenes;
-mod cob_using;
 
 pub use cob_commands::*;
 pub use cob_defs::*;
 pub use cob_import::*;
 pub use cob_manifest::*;
 pub use cob_scenes::*;
-pub use cob_using::*;