+++ src/loading/extract/mod.rs
@@ -4,7 +4,6 @@ mod extract_defs;
 mod extract_import;
 mod extract_manifest;
 mod extract_scenes;
-mod extract_using;
 mod reflected_loadable;
 mod utils;
 
@@ -14,6 +13,5 @@ pub(self) use extract_defs::*;
 pub(self) use extract_import::*;
 pub(self) use extract_manifest::*;
 pub(self) use extract_scenes::*;
-pub(self) use extract_using::*;
 pub(crate) use reflected_loadable::*;
 pub(self) use utils::*;