+++ src/loading/extract/utils.rs
@@ -1,5 +1,4 @@
 use std::any::TypeId;
-use std::collections::HashMap;
 use std::sync::Arc;
 
 use bevy::reflect::serde::TypedReflectDeserializer;
@@ -15,34 +14,30 @@ pub(super) fn get_loadable_meta<'a>(
     file: &CobFile,
     current_path: &ScenePath,
     short_name: &str,
-    name_shortcuts: &mut HashMap<&'static str, &'static str>,
+    loadables: &LoadableRegistry,
 ) -> Option<(&'static str, &'static str, TypeId, TypedReflectDeserializer<'a>)>
 {
-    // Check if we already have this mapping.
-    let mut found_mapping = false;
-    let registration = match name_shortcuts.get(short_name) {
-        Some(long_name) => {
-            found_mapping = true;
-            type_registry.get_with_type_path(long_name)
+    // Look up the registration.
+    let registration = match loadables.get_type_id(short_name) {
+        Some(type_id) => type_registry.get(type_id),
+        None => {
+            tracing::warn!("failed getting type id for loadable {} at {:?} in {:?}; no loadable with this name was \
+                registered in the app",
+                short_name, current_path, file);
+            return None;
         }
-        None => type_registry.get_with_short_type_path(short_name),
     };
 
     // Look up the longname
     let Some(registration) = registration else {
-        tracing::error!("failed getting long type name for {:?} at {:?} in {:?}; if the type is ambiguous because \
-            there are multiple types with this short name, add its long name to the cobweb asset file's 'using' section",
+        tracing::error!("failed getting type registration for {} at {:?} in {:?}; type was not registered in the app \
+            (this is a bug)",
             short_name, current_path, file);
         return None;
     };
 
     let short_name = registration.type_info().type_path_table().short_path(); //get static version
     let long_name = registration.type_info().type_path_table().path();
-
-    // Save this mapping for later.
-    if !found_mapping {
-        name_shortcuts.insert(short_name, long_name);
-    }
 
     // Deserializer
     let deserializer = TypedReflectDeserializer::new(registration, type_registry);